Allow discovery of devices on the USB-C port.
Storage devices on the USB-C port were not being discovered. Only an Android
phone or an USB-C to HDMI adapter would trigger a dmesg response when plugged.

With the included patch, USB storage devices on the USB-C port via a type-C to
type-A adapter are immediately discovered and assigned to the right bus(480M
or 5000M). The same observation is done for a USB-3 hub.